+++ src/components/feature/LandingPage/HeroSection/HeroCtaButton.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 import { Upload } from 'lucide-react';
-import { Button } from '../../common/Button';
+import { Button } from '../../../common/Button';
 
 export function HeroCtaButton() {
   const navigate = useNavigate();
